requirements.txt
flask==2.0.1
flask-cors==3.0.10
pydantic==1.8.2
faiss-cpu==1.7.2
numpy==1.21.2
requests==2.26.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0
google-generativeai==0.3.0
markitdown==0.1.0
mcp-server==0.1.0
tqdm==4.62.3
Pillow==8.3.2 